Rename EventDescriptor type
