+++ src/script-handlers/classPropHandler.ts
@@ -1,9 +1,8 @@
 import * as bt from '@babel/types'
 import { NodePath } from 'ast-types'
-import { Documentation } from '../Documentation'
-import { BlockTag } from '../utils/blockTags'
+import { BlockTag, DocBlockTags, Documentation } from '../Documentation'
 import getDocblock from '../utils/getDocblock'
-import getDoclets, { DocBlockTags } from '../utils/getDoclets'
+import getDoclets from '../utils/getDoclets'
 import getTypeFromAnnotation from '../utils/getTypeFromAnnotation'
 import transformTagsIntoObject from '../utils/transformTagsIntoObject'
 import { describeDefault, describeRequired } from './propHandler'
