+++ src/utils/getTypeFromAnnotation.ts
@@ -1,5 +1,5 @@
 import * as bt from '@babel/types'
-import { ParamType } from './getDoclets'
+import { ParamType } from '../Documentation'
 
 export default function getTypeFromAnnotation(
   typeNode: bt.TypeAnnotation | bt.TSTypeAnnotation | bt.Noop | null,
